Ready handler migrated
- moved ready event handler from main to separate file
- added ready event handler to event index
- made main tests work with 'once' handlers too
