+++ src/main.ts
@@ -1,13 +1,8 @@
 import 'source-map-support/register';
 import 'dotenv/config';
-import { ActivityType, Client, GatewayIntentBits, Partials } from 'discord.js';
-import { appVersion } from './constants/meta';
-import { deployCommands } from './helpers/actions/deployCommands';
-import { registerEventHandlers } from './events'; // TODO update tests to reflect new event handlers
-import { parseArgs } from './helpers/parseArgs';
-import { revokeCommands } from './helpers/actions/revokeCommands';
-import { verifyCommandDeployments } from './helpers/actions/verifyCommandDeployments';
+import { Client, GatewayIntentBits, Partials } from 'discord.js';
 
+import { registerEventHandlers } from './events'; // TODO update tests to reflect new event handlers
 import { getLogger } from './logger';
 const logger = getLogger();
 
@@ -15,6 +10,9 @@ const logger = getLogger();
 // none of this setup would be testable :P
 
 export async function _main(): Promise<void> {
+	logger.info('*Yawn* Good morning!');
+
+	// Set up the client
 	const client = new Client({
 		intents: [
 			GatewayIntentBits.Guilds,
@@ -30,42 +28,10 @@ export async function _main(): Promise<void> {
 		},
 	});
 
-	const args = parseArgs();
-
-	logger.info('*Yawn* Good morning!');
-
-	client.on('ready', async client => {
-		logger.info(`Starting ${client.user.username} v${appVersion}...`);
-
-		// If we're only here to deploy commands, do that and then exit
-		if (args.deploy) {
-			await deployCommands(client);
-			client.destroy();
-			return;
-		}
-
-		// If we're only here to revoke commands, do that and then exit
-		if (args.revoke) {
-			await revokeCommands(client);
-			client.destroy();
-			return;
-		}
-
-		// Sanity check for commands
-		logger.info('Verifying command deployments...');
-		await verifyCommandDeployments(client);
-
-		// Let users know where to go for info
-		client.user.setActivity({
-			type: ActivityType.Playing,
-			name: '/help for info',
-		});
-
-		logger.info('Ready!');
-	});
-
+	// Register all the event handlers for the client
 	registerEventHandlers(client);
 
+	// Login
 	try {
 		await client.login(process.env['DISCORD_TOKEN']);
 	} catch (error) {
